+++ CHANGELOG.md
@@ -1,7 +1,17 @@
 Changelog
 =========
-v0.1.1
+
+**v0.2.0**
+
+- Add configuration options to the panel view:
+    * Minimum/maximum width
+    * Show progressbar
+    * Hide disabled controls
+
+**v0.1.1**
+
 - Fix sources tabbar continuous update
 
-v0.1.0
+**v0.1.0**
+
 - Modified version with tabbar and compact view for panels